# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface ApiError {
  error: string;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  beef: "beef",
  pork: "pork",
  chicken: "chicken",
  lamb: "lamb",
  veal: "veal",
  game: "game",
} as const;

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  cut: string;
  origin: string;
  grade: string;
  pricePerKgCents: number;
  stockKg: number;
  unit: string;
  imageUrl: string;
  description: string;
  createdAt: string;
}

export type ProductInputCategory =
  (typeof ProductInputCategory)[keyof typeof ProductInputCategory];

export const ProductInputCategory = {
  beef: "beef",
  pork: "pork",
  chicken: "chicken",
  lamb: "lamb",
  veal: "veal",
  game: "game",
} as const;

export interface ProductInput {
  name: string;
  category: ProductInputCategory;
  cut: string;
  origin: string;
  grade: string;
  pricePerKgCents: number;
  stockKg: number;
  unit: string;
  imageUrl: string;
  description: string;
}

export type CustomerAccountType =
  (typeof CustomerAccountType)[keyof typeof CustomerAccountType];

export const CustomerAccountType = {
  restaurant: "restaurant",
  butcher: "butcher",
  hotel: "hotel",
  retailer: "retailer",
  distributor: "distributor",
} as const;

export interface Customer {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  accountType: CustomerAccountType;
  creditLimitCents: number;
  notes: string;
  createdAt: string;
}

export type CustomerInputAccountType =
  (typeof CustomerInputAccountType)[keyof typeof CustomerInputAccountType];

export const CustomerInputAccountType = {
  restaurant: "restaurant",
  butcher: "butcher",
  hotel: "hotel",
  retailer: "retailer",
  distributor: "distributor",
} as const;

export interface CustomerInput {
  businessName: string;
  contactName: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  accountType: CustomerInputAccountType;
  creditLimitCents: number;
  notes: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantityKg: number;
  pricePerKgCents: number;
  lineTotalCents: number;
}

export interface OrderItemInput {
  productId: number;
  quantityKg: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  status: OrderStatus;
  deliveryDate: string;
  totalCents: number;
  notes: string;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  customerId: number;
  deliveryDate: string;
  notes: string;
  items: OrderItemInput[];
}

export interface DashboardSummary {
  totalRevenueCents: number;
  ordersThisMonth: number;
  activeCustomers: number;
  pendingOrders: number;
  totalProducts: number;
  lowStockCount: number;
  revenueChangePct: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  category: string;
  totalKg: number;
  totalRevenueCents: number;
}

export interface CategorySales {
  category: string;
  totalKg: number;
  totalRevenueCents: number;
  orderCount: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
};

export type ListOrdersParams = {
  status?: string;
  customerId?: number;
};