# lib/api-zod/src/generated/types/categorySales.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySales {
  category: string;
  totalKg: number;
  totalRevenueCents: number;
  orderCount: number;
}