# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerAccountType } from "./customerAccountType";

export interface Customer {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  accountType: CustomerAccountType;
  creditLimitCents: number;
  notes: string;
  createdAt: Date;
}