# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRevenueCents: number;
  ordersThisMonth: number;
  activeCustomers: number;
  pendingOrders: number;
  totalProducts: number;
  lowStockCount: number;
  revenueChangePct: number;
}