# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  status: OrderStatus;
  deliveryDate: Date;
  totalCents: number;
  notes: string;
  items: OrderItem[];
  createdAt: Date;
}