# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * B2B meat supply API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  cut: string;
  origin: string;
  grade: string;
  pricePerKgCents: number;
  stockKg: number;
  unit: string;
  imageUrl: string;
  description: string;
  createdAt: Date;
}